Add duplicate name notification
Add a notification for if the user enters a name that
is already on the tree.

TODO: When the queue is added for names that the
moderator needs to modify, both the list of names on the
tree and the list of names in the queue need to be
checked for a duplicate name.